requirements.txt
langchain
langchain-openai
langchain-mcp-adapters
mcp
fastmcp
python-dotenv